fix(flags): Phase 1D-1 review fixes — per-cat routing + pagination + Long? alignment
Adresse les 5 issues bloquantes du verdict Codex sur PR #112.

- **P112-B1, P112-B2** — bascule vers la voie REST per-cat via
  `HfrApiClient.getCategoryFlagTopics(...)` au lieu de la voie globale
  `getFlagTopics(...)` dont l'enveloppe groupée par catégorie n'est pas
  capturée. `DefaultFlagRepository` injecte désormais `ForumRepository` pour
  récupérer la liste des catégories publiques et fan-out parallèle (`async`/
  `awaitAll`) sur chaque cat. Tests adaptés pour mocker `ForumRepository` et
  pointer sur la fixture capturée `rest_cat23_participated.json`. Les payloads
  synthétiques contractuels inline sont retirés ; les cas défensifs
  (pagination, échec) restent synthétiques mais explicitement marqués.
- **P112-B3** — pagination per-cat : le repository walk les pages tant que
  `page * results_per_page < results_count`. Test synthétique multi-page.
- **P112-B4, Missed-1** — `RestTopic.lastPostReadId: Int?` → `Long?` pour
  absorber un `numreponse` HFR au-delà de `Int.MAX_VALUE` sans crash de
  désérialisation. `TopicSummary.lastPostReadId` aligné en `Long?` ;
  `RedfaceNavigation` ajoute le narrowing borné côté category route. Le mapper
  flags `dto.lastPostReadId?.toLong()` devient `dto.lastPostReadId`.
- **P112-B5, Missed-2** — KDoc `FlagRepository` rafraîchie Phase 1B.4 →
  Phase 1D-1. `protocol-hfr.md`, `architecture.md` et `rest_cat23_participated.
  source.txt` actent que la voie consommée est per-cat ; le global REST reste
  routé en plumbing mais non consommé tant qu'aucune fixture live n'existe.
  CHANGELOG ajoute une section "Limites connues".

Refs #110

+++ core/model/src/main/kotlin/fr/forumhfr/redface2/core/model/TopicSummary.kt
@@ -44,8 +44,12 @@ data class TopicSummary(
      * Authenticated-only field, mirrors REST `last_post_read_id` — the id of
      * the last post the user actually read (anchor for a "scroll to last read"
      * deep link). Distinct from "first unread", which is one post further.
+     * Stored as `Long?` so the DTO/mapper chain handles HFR `numreponse` values
+     * past `Int.MAX_VALUE` (~2.1 G) without a deserialisation crash. The
+     * navigation layer narrows back to `Int` for [TopicRoute.scrollTo] when the
+     * value fits.
      */
-    val lastPostReadId: Int?,
+    val lastPostReadId: Long?,
     /**
      * Authenticated-only flag bucket inferred from REST `flag_owntopic` :
      * - `1 → FlagType.CYAN`     (« Mes sujets » — sujets participés)
